+++ src/platform/packages/private/kbn-split-button/.storybook/main.js
@@ -7,4 +7,4 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-export * from './split_button';
+module.exports = require('@kbn/storybook').defaultConfig;
