+++ src/platform/packages/private/kbn-split-button/README.md
@@ -1,0 +1,3 @@
+# @kbn/split-button
+
+Empty package generated by @kbn/generate
