+++ src/platform/packages/private/kbn-split-button/jest.config.js
@@ -7,19 +7,8 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import { defaultConfig } from '@kbn/storybook';
-import { merge as webpackMerge } from 'webpack-merge';
-import { resolve } from 'path';
-
-const mockConfig = {
-  resolve: {
-    alias: {
-      '@kbn/esql/public': resolve(__dirname, '../public/mocks/esql_editor.tsx'),
-    },
-  },
-};
-
 module.exports = {
-  ...defaultConfig,
-  webpackFinal: (config) => webpackMerge(config, mockConfig),
+  preset: '@kbn/test',
+  rootDir: '../../../../..',
+  roots: ['<rootDir>/src/platform/packages/private/kbn-split-button'],
 };
