+++ src/platform/packages/private/kbn-split-button/kibana.jsonc
@@ -1,0 +1,7 @@
+{
+  "type": "shared-common",
+  "id": "@kbn/split-button",
+  "owner": "@elastic/kibana-data-discovery",
+  "group": "platform",
+  "visibility": "private"
+}
